requirements.txt
# Deep Learning Framework
torch>=1.9.0
torchvision>=0.10.0

# Data Processing
pandas>=1.3.0
numpy>=1.19.0
scikit-learn>=0.24.0

# Text Processing
nltk>=3.6.0
tqdm>=4.61.0

# Medical Text Processing
gensim>=4.0.0  # For word embeddings
transformers>=4.9.0  # If using BERT or similar models

# Visualization and Logging
matplotlib>=3.4.0
tensorboard>=2.6.0

# Utilities
PyYAML>=5.4.0  # For config files
python-dotenv>=0.19.0  # For environment variables

# Optional but recommended
jupyter>=1.0.0  # For notebook experimentation 